lisotto/requirements.txt
numpy==1.15.4
PyYAML==3.13
tensorflow-gpu==1.13.0
beautifultable==0.5.3
